pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "probabilistic-epf-of-eu-bidding-zones"
version = "1.0.0"
description = "Probabilistic EPF experiments (Normalizing Flow, NHITS+QRA, Moirai, Chronos/ ChronosX)"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12,<3.13"

authors = [
  { name = "Jan Lettner", email = "yw8119@partner.kit.edu" },
  { name = "Hadeer El Ashhab", email = "hadeer.elashhab@kit.edu" }
]

license = { file = "LICENSE" }
keywords = ["time-series", "forecasting", "probabilistic", "energy", "transfer-learning", "properscoring", "electricity"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  # keep this SMALL: only what *all* stacks need
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "rich>=13",
  "matplotlib~=3.10",
  "tensorboard",
  "statsmodels",
  "pyarrow"
]

[project.urls]
Public-Repository = "https://github.com/MagJ1/probabilistic-epf-of-eu-bidding-zones"
Dracos-Group = "https://www.iai.kit.edu/fachgebiete_4102.php"



# ---- extras: one per model stack ----
[project.optional-dependencies]
moirai = [
  "uni2ts @ git+https://github.com/SalesforceAIResearch/uni2ts.git@8d2a08d",
  "torch>=2.1,<2.5",
  "lightning>=2.0",
  "gluonts~=0.14.3",
  "scipy~=1.11.3",
  "einops==0.7.*",
  "jaxtyping~=0.2.24",
  "python-dotenv==1.0.0",
  "orjson",
  "multiprocess",
  "huggingface_hub>=0.23.0",
  "safetensors",
  "datasets~=2.17.1",
  "jax[cpu]",
  "joblib"
]

chronosx = [
  "chronosx @ git+https://github.com/amazon-science/chronos-forecasting.git@2b52bfc",
  "transformers==4.47.1",
  "accelerate==0.34.2",
  "datasetsforecast==1.0.0",
  "fev==0.5.0",
  "lightning>=2.0",
  "joblib"
]

normalizing_flows = [
  "joblib==1.5.3",
  "normflows==1.7.3",
  "numpy>=1.26,<3",
  "optuna==4.6.0",
  "pandas==2.3.3",
  "properscoring==0.1",
  "lightning>=2.0",
  "scikit-learn==1.8.0",
  "scipy>=1.11,<2",
  "torch>=2.1,<2.5",
]

nhits_qra = [
  "cvxpy==1.7.5",
  "joblib==1.5.3",
  "neuralforecast==3.1.2",
  "numpy>=1.26,<3",
  "optuna==4.6.0",
  "pandas==2.3.3",
  "lightning>=2.0",
  "scikit-learn==1.8.0",
  "scipy>=1.11,<2",
  "torch>=2.1,<2.5",
  "tqdm==4.67.1",
]

dev = [
  "pytest>=8",
  "ruff>=0.6",
  "pre-commit>=4",
  "pytest-cov"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"models.moirai.conf" = ["**/*.yaml", "**/*.yml"]
"models.chronosx.conf" = ["**/*.yaml", "**/*.yml"]
"models.normalizing_flows.conf" = ["**/*.yaml", "**/*.yml"]
"models.nhits_qra.conf" = ["**/*.yaml", "**/*.yml"]